Move data_interface classes/funcs that need GUI to new gui file
